+++ CasaEcologieSysInfo/ResStockMatieresPremiere.cs
@@ -24,7 +24,7 @@ namespace CasaEcologieSysInfo
     
         public int CodeMatierePremiere { get; set; }
         public string NomMatiere { get; set; }
-        public int StockMatiere { get; set; }
+        public float StockMatiere { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<EveReceptionMatieresPremiere> EveReceptionMatieresPremieres { get; set; }
